Begin conversion to typescript
