--- frontend/busk_app_frontend/src/Error.js
@@ -1,20 +1,0 @@
-import "./Error.css";
-
-/** Renders an error message
- * 
- * State:
- *  - none
- * 
- * Props:
- *  - Array of errror messages
- * 
- */
-function Error({ messages }) {
-    return (
-        <div className="Error">
-            {messages.map((message, idx) => <p key={idx}>{message}</p>)}
-        </div>
-    );
-}
-
-export default Error;
