+++ frontend/src/Home.tsx
@@ -1,0 +1,23 @@
+import React from "react";
+
+// import "./HomePage.css";
+/** Renders HomePage
+ *
+ * Props: none
+ * State: none
+ *
+ * Routes --> HomePage
+ */
+
+function Home() {
+  return (
+    <div className="Homepage">
+      <div className="container text-center">
+        <h1 className="mb-4 fw-bold">Welcome To Busk!</h1>
+        <p className="lead">Placeholder!</p>
+      </div>
+    </div>
+  );
+}
+
+export default Home;
